+++ src/domains/followups/index.ts
@@ -1,0 +1,3 @@
+export * from './hooks/useFollowUpsViewModel';
+export * from './types';
+export * from './helpers';