+++ src/domains/projects/index.ts
@@ -1,0 +1,3 @@
+export * from './hooks/useProjectsViewModel';
+export * from './types';
+export * from './helpers';