+++ src/domains/tasks/index.ts
@@ -1,0 +1,3 @@
+export * from './hooks/useTasksViewModel';
+export * from './types';
+export * from './helpers';